# openapi.yaml
openapi: 3.0.3
info:
  title: LIS Labo-Anapath API
  description: API REST de gestion de laboratoire d'anatomopathologie
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Développement local
  - url: https://api.labo-anapath.com/api/v1
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /consultations:
    get:
      tags: [Consultations]
      summary: Liste des consultations
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - name: patientId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationPageApiResponse'
    post:
      tags: [Consultations]
      summary: Créer une consultation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsultationRequest'
      responses:
        '201':
          description: Créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationApiResponse'
  /consultations/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags: [Consultations]
      summary: Détail
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationApiResponse'
    put:
      tags: [Consultations]
      summary: Modifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsultationRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsultationApiResponse'
    delete:
      tags: [Consultations]
      summary: Supprimer
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'

  # ── TYPE CONSULTATIONS ───────────────────────────────────────────────────────
  /type-consultations:
    get:
      tags: [Type Consultations]
      summary: Types de consultation
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeConsultationPageApiResponse'
    post:
      tags: [Type Consultations]
      summary: Créer un type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimpleNameRequest'
      responses:
        '201':
          description: Créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeConsultationApiResponse'
  /type-consultations/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags: [Type Consultations]
      summary: Détail
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeConsultationApiResponse'
    put:
      tags: [Type Consultations]
      summary: Modifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimpleNameRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypeConsultationApiResponse'
    delete:
      tags: [Type Consultations]
      summary: Supprimer
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'

  # ── CONTRACTS ────────────────────────────────────────────────────────────────
  /contracts:
    get:
      tags: [Contracts]
      summary: Liste des contrats
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - name: hospitalId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContratPageApiResponse'
    post:
      tags: [Contracts]
      summary: Créer un contrat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContratRequest'
      responses:
        '201':
          description: Créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContratApiResponse'
  /contracts/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags: [Contracts]
      summary: Détail d'un contrat
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContratApiResponse'
    put:
      tags: [Contracts]
      summary: Modifier un contrat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContratRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContratApiResponse'
    delete:
      tags: [Contracts]
      summary: Supprimer un contrat
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'

  # ── INVOICES ─────────────────────────────────────────────────────────────────
...
  /employees/{employeeId}/contrats/{id}:
    parameters:
      - name: employeeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/id'
    put:
      tags: [Employees]
      summary: Modifier un contrat RH
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeContratRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeContratApiResponse'
    delete:
      tags: [Employees]
      summary: Supprimer un contrat RH
      responses:
        '200':
          $ref: '#/components/responses/EmptySuccess'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── REUSABLE PARAMETERS ──────────────────────────────────────────────────────
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
    size:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  responses:
    EmptySuccess:
      description: Opération réussie
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EmptyApiResponse'
...
  schemas:
...
    PageMeta:
      type: object
      properties:
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          format: int64
          example: 100
        totalPages:
          type: integer
          example: 5
        last:
          type: boolean
    EmptyApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: OK
        data:
          nullable: true
        timestamp:
          type: string
          format: date-time
...
    SimpleNameRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Hématologie
...
    ConsultationRequest:
      type: object
      required: [patientId, date]
      properties:
        patientId:
          type: string
          format: uuid
        doctorId:
          type: string
          format: uuid
        typeConsultationId:
          type: string
          format: uuid
        notes:
          type: string
        date:
          type: string
          format: date
    ConsultationResponse:
      type: object
      properties:
        id: {type: string, format: uuid}
        patientId: {type: string, format: uuid}
        patientFirstName: {type: string}
        patientLastName: {type: string}
        doctorId: {type: string, format: uuid}
        doctorLastName: {type: string}
        typeConsultationId: {type: string, format: uuid}
        typeConsultationName: {type: string}
        notes: {type: string}
        date: {type: string, format: date}
        branchId: {type: string, format: uuid}
        createdAt: {type: string, format: date-time}
    ConsultationApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          $ref: '#/components/schemas/ConsultationResponse'
        timestamp: {type: string, format: date-time}
    ConsultationPageApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          allOf:
            - $ref: '#/components/schemas/PageMeta'
            - type: object
              properties:
                content:
                  type: array
                  items:
                    $ref: '#/components/schemas/ConsultationResponse'
        timestamp: {type: string, format: date-time}
    TypeConsultationResponse:
      type: object
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        branchId: {type: string, format: uuid}
    TypeConsultationApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          $ref: '#/components/schemas/TypeConsultationResponse'
        timestamp: {type: string, format: date-time}
    TypeConsultationPageApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          allOf:
            - $ref: '#/components/schemas/PageMeta'
            - type: object
              properties:
                content:
                  type: array
                  items:
                    $ref: '#/components/schemas/TypeConsultationResponse'
        timestamp: {type: string, format: date-time}

    # -- Contracts ------------------------------------------------------------
    ContratDetailRequest:
      type: object
      properties:
        labTestId:
          type: string
          format: uuid
        price:
          type: number
          format: double
    ContratDetailResponse:
      type: object
      properties:
        labTestId: {type: string, format: uuid}
        labTestName: {type: string}
        price: {type: number, format: double}
    ContratRequest:
      type: object
      required: [startDate]
      properties:
        hospitalId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        nbrTests:
          type: integer
        status:
          type: string
          default: ACTIVE
        details:
          type: array
          items:
            $ref: '#/components/schemas/ContratDetailRequest'
    ContratResponse:
      type: object
      properties:
        id: {type: string, format: uuid}
        hospitalId: {type: string, format: uuid}
        hospitalName: {type: string}
        nbrTests: {type: integer}
        startDate: {type: string, format: date}
        endDate: {type: string, format: date, nullable: true}
        status: {type: string}
        details:
          type: array
          items:
            $ref: '#/components/schemas/ContratDetailResponse'
        branchId: {type: string, format: uuid}
        createdAt: {type: string, format: date-time}
    ContratApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          $ref: '#/components/schemas/ContratResponse'
        timestamp: {type: string, format: date-time}
    ContratPageApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          allOf:
            - $ref: '#/components/schemas/PageMeta'
            - type: object
              properties:
                content:
                  type: array
                  items:
                    $ref: '#/components/schemas/ContratResponse'
        timestamp: {type: string, format: date-time}

    # -- Finance --------------------------------------------------------------
...
    EmployeeContratRequest:
      type: object
      required: [startDate, type, salary]
      properties:
        startDate: {type: string, format: date}
        endDate: {type: string, format: date}
        type: {type: string, example: CDI}
        salary: {type: number, format: double}
    EmployeeContratResponse:
      type: object
      properties:
        id: {type: string, format: uuid}
        startDate: {type: string, format: date}
        endDate: {type: string, format: date, nullable: true}
        type: {type: string}
        salary: {type: number, format: double}
        createdAt: {type: string, format: date-time}
    EmployeeContratApiResponse:
      type: object
      properties:
        success: {type: boolean}
        message: {type: string}
        data:
          $ref: '#/components/schemas/EmployeeContratResponse'
        timestamp: {type: string, format: date-time}
...